Add filters for documents
