+++ deployments/drf_views.py
@@ -1,7 +1,6 @@
-from django.db.models import Q
 from rest_framework.authentication import TokenAuthentication
 from rest_framework.permissions import IsAuthenticated
-from django_filters import Filter, rest_framework as filters
+from django_filters import rest_framework as filters
 from rest_framework import viewsets
 from .models import (
     ERUOwner,
@@ -14,6 +13,7 @@ from .models import (
     PartnerSocietyDeployment,
 )
 from api.models import Country
+from api.view_filters import ListFilter
 from .serializers import (
     ERUOwnerSerializer,
     ERUSerializer,
@@ -25,24 +25,6 @@ from .serializers import (
     PartnerDeploymentSerializer,
 )
 
-# https://github.com/carltongibson/django-filter/issues/137#issuecomment-127073878
-class ListFilter(Filter):
-    def __init__(self, filter_value=lambda x: x, **kwargs):
-        super(ListFilter, self).__init__(**kwargs)
-        self.filter_value_fn = filter_value
-
-    def sanitize(self, value_list):
-        return [v for v in value_list if v != u'']
-
-    def filter(self, qs, value):
-        values = value.split(u",")
-        values = self.sanitize(values)
-        values = map(self.filter_value_fn, values)
-        f = Q()
-        for v in values:
-            kwargs = {self.name: v}
-            f = f|Q(**kwargs)
-        return qs.filter(f)
 
 class ERUOwnerViewset(viewsets.ReadOnlyModelViewSet):
     authentication_classes = (TokenAuthentication,)
